# specs/001-backend-mvp/contracts/api.yaml
openapi: 3.0.3
info:
  title: MiniProntuario API
  version: 1.0.0
  description: Backend MVP API for MiniProntuario Odontologico
paths:
  /auth/register:
    post:
      summary: Register a new dentist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Created successfully
        '400':
          description: Bad request (e.g. duplicate CPF or Email)

  /auth/login:
    post:
      summary: Authenticate dentist and obtain JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Unauthorized

  /patients:
    post:
      summary: Register a new patient
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientRequest'
      responses:
        '201':
          description: Patient created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
        '400':
          description: Validation error or Duplicate CPF for this dentist

    get:
      summary: List all patients for the authenticated dentist
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PatientResponse'

  /patients/{id}:
    get:
      summary: Get a specific patient
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Patient data
        '404':
          description: Patient not found or belongs to another dentist

    delete:
      summary: Soft delete a patient
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted successfully

  /patients/{patientId}/procedures:
    post:
      summary: Add a procedure to a patient
      security:
        - bearerAuth: []
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcedureRequest'
      responses:
        '201':
          description: Procedure created
        '400':
          description: Validation error (e.g., date in future)

    get:
      summary: List all procedures for a patient
      security:
        - bearerAuth: []
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of procedures

  /procedures/{id}:
    put:
      summary: Update a procedure (within 24h window)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcedureRequest'
      responses:
        '200':
          description: Procedure updated
        '403':
          description: Forbidden (past 24h edit window)
        '404':
          description: Procedure not found

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    RegisterRequest:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        password: { type: string }
        cpf: { type: string }
        cro: { type: string }
        phone: { type: string }
    LoginRequest:
      type: object
      properties:
        email: { type: string }
        password: { type: string }
    AuthResponse:
      type: object
      properties:
        token: { type: string }
    PatientRequest:
      type: object
      properties:
        name: { type: string }
        cpf: { type: string }
        birthDate: { type: string, format: date }
        phone: { type: string }
        allergies: { type: string }
        systemicDiseases: { type: string }
    PatientResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        cpf: { type: string }
        birthDate: { type: string, format: date }
        phone: { type: string }
        allergies: { type: string }
        systemicDiseases: { type: string }
        createdAt: { type: string, format: date-time }
    ProcedureRequest:
      type: object
      properties:
        date: { type: string, format: date }
        description: { type: string }
        tooth: { type: string }
        notes: { type: string }